+++ src/ui.js
@@ -1529,6 +1529,8 @@ var KanvazUI_Extended = (function() {
           ['Shift+Arrow','Nudge 10px'],
           ['Delete',     'Delete card'],
           ['Ctrl+D',     'Duplicate'],
+          ['Ctrl+G',     'Group selected cards (2+)'],
+          ['Ctrl+Shift+G', 'Ungroup'],
           ['P',          'Pin / unpin'],
           ['A',          'Annotate'],
           ['H',          'Hide annotations'],